+++ ascend-sdk/src/main/java/ai/evolv/AscendAction.java
@@ -1,4 +1,4 @@
-package ai.evolv.ascend.android;
+package ai.evolv;
 
 public interface AscendAction<T> {
