+++ ascend-sdk/src/main/java/ai/evolv/AscendAllocationStore.java
@@ -1,4 +1,4 @@
-package ai.evolv.ascend.android;
+package ai.evolv;
 
 import com.google.gson.JsonArray;
