+++ ascend-sdk/src/main/java/ai/evolv/AscendConfig.java
@@ -1,4 +1,4 @@
-package ai.evolv.ascend.android;
+package ai.evolv;
 
 public class AscendConfig {
