+++ ascend-sdk/src/main/java/ai/evolv/AscendParticipant.java
@@ -1,4 +1,4 @@
-package ai.evolv.ascend.android;
+package ai.evolv;
 
 import java.util.UUID;
