+++ ascend-sdk/src/main/java/ai/evolv/exceptions/AscendAllocationException.java
@@ -1,4 +1,4 @@
-package ai.evolv.ascend.android.exceptions;
+package ai.evolv.exceptions;
 
 public class AscendAllocationException extends Exception {
     public AscendAllocationException(String errorMessage) {
