+++ ascend-sdk/src/test/java/ai/evolv/AscendParticipantTest.java
@@ -1,12 +1,8 @@
-package ai.evolv.ascend.android;
-
-import android.support.test.runner.AndroidJUnit4;
+package ai.evolv;
 
 import org.junit.Assert;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
-@RunWith(AndroidJUnit4.class)
 public class AscendParticipantTest {
 
     @Test
