+++ ascend-sdk/src/test/java/ai/evolv/EventEmitterTest.java
@@ -1,12 +1,9 @@
-package ai.evolv.ascend.android;
-
-import android.support.test.runner.AndroidJUnit4;
+package ai.evolv;
 
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
 import static org.mockito.Mockito.*;
 import org.mockito.Mock;
@@ -14,7 +11,6 @@ import org.mockito.MockitoAnnotations;
 
 import okhttp3.HttpUrl;
 
-@RunWith(AndroidJUnit4.class)
 public class EventEmitterTest {
 
     private static final String environmentId = "test_12345";
